+++ public/index.html
@@ -48,7 +48,7 @@
     </div>
     <script type="module" src="/js/main.js"></script>
     <script type="module" src="/js/controllers/music-controller.js"></script>
-    <script type="module" src="/js/player-feedback.js"></script>
-    <script type="module" src="/js/games-state-listener.js"></script>
+    <script type="module" src="/js/components/player-feedback.js"></script>
+    <script type="module" src="/js/listeners/games-state-listener.js"></script>
   </body>
 </html>
